# c_openapi.yaml
openapi: 3.1.0
info:
  title: 住民記録システム Web版 API
  version: 0.1.0
  description: |
    住民記録システム標準仕様書 第6.1版（令和8年3月25日版 / 自治体システム等標準化検討会）に準拠した
    REST API の設計仕様（ドラフト）。ChromeOS Flex を含む Chromium 系ブラウザからの利用を想定。

    - すべての永続化操作は監査ログに記録される。
    - すべての応答は UTF-8 / application/json。日付は ISO 8601。
    - 認可は OAuth2 / OIDC のスコープ + 属性ベース（ロール + 項目別マスク）。
servers:
  - url: https://{tenant}.juki.example.go.jp/api/v1
    variables:
      tenant: { default: city01 }
...
security:
  - oidc: [residents.read, residents.write, transactions.write, certificates.issue, admin]
paths:
...
  /transactions/in:
    post:
      tags: [transactions]
      summary: 転入届の受付・反映
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MoveInReq' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } }
  /transactions/out:
    post:
      tags: [transactions]
      summary: 転出届
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MoveOutReq' } } } }
      responses:
        '201':
          description: Created（転出証明書発行情報を含む）
          content: { application/json: { schema: { $ref: '#/components/schemas/TransactionWithCertificate' } } }
  /transactions/move:
    post:
      tags: [transactions]
      summary: 転居届
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MoveInternalReq' } } } }
      responses: { '201': { description: Created } }
  /transactions/household:
    post:
      tags: [transactions]
      summary: 世帯変更（世帯主変更・分離・合併）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/HouseholdChangeReq' } } } }
      responses: { '201': { description: Created } }
...
  /transactions/official:
    post:
      tags: [transactions]
      summary: 職権異動 起票
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OfficialTxReq' } } } }
      responses: { '201': { description: Created } }
  /transactions/{txId}/approve:
    post:
      tags: [transactions]
      summary: 職権異動 決裁（承認/差戻/却下）
      parameters: [ { name: txId, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ApprovalReq' } } }
      responses: { '200': { description: OK } }
  /transactions/cancel:
    post:
      tags: [transactions]
      summary: 異動取消（標準仕様 4.6）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CancelReq' } } } }
      responses: { '201': { description: Created } }

  # 住民票コード / 個人番号 / 外国人在留
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.example.go.jp/.well-known/openid-configuration
...
  schemas:
...
    ForeignerInfo:
      type: object
      properties:
        residenceStatus:        { type: string }
        residencePeriodEnd:     { type: string, format: date }
        passportNo:             { type: string }
        nationalityFull:        { type: string }
        aliasKanji:             { type: string, description: 通称 }
        specialPermanentResident: { type: boolean }
...
    Transaction:
      type: object
      properties:
        transactionId:        { type: string }
        residentId:           { type: string }
        householdId:          { type: string }
        typeCode:             { type: string, description: 'IN/OUT/MOVE/HH/BIRTH/DEATH/OFFICIAL/CANCEL ...' }
        reasonCode:           { type: string }
        eventDate:            { type: string, format: date }
        processedDate:        { type: string, format: date }
        receiverOffice:       { type: string }
        status:               { type: string, enum: [DRAFT, REVIEW, APPROVED, APPLIED, CANCELLED] }
        parentTransactionId:  { type: string, nullable: true }
        items:                { type: array, items: { $ref: '#/components/schemas/TransactionItem' } }
    TransactionItem:
      type: object
      properties:
        field:       { type: string }
        valueBefore: { type: string, nullable: true }
        valueAfter:  { type: string, nullable: true }
    MoveInReq:
      type: object
      required: [eventDate, members]
      properties:
        eventDate:           { type: string, format: date }
        previousAddress:     { type: string }
        moveOutCertNo:       { type: string }
        members:             { type: array, items: { $ref: '#/components/schemas/MoveMember' } }
    MoveMember:
      type: object
      properties:
        familyNameKanji: { type: string }
        givenNameKanji:  { type: string }
        familyNameKana:  { type: string }
        givenNameKana:   { type: string }
        birthDate:       { type: string, format: date }
        sex:             { type: string, enum: [M, F, U] }
        relationToHead:  { type: string }
        myNumber:        { type: string }
        foreigner:       { $ref: '#/components/schemas/ForeignerInfo' }
    MoveOutReq:
      type: object
      required: [eventDate, members, newAddress]
      properties:
        eventDate:    { type: string, format: date }
        newAddress:   { type: string }
        newAddressCode: { type: string }
        reasonCode:   { type: string }
        members:      { type: array, items: { type: string, description: residentId } }
    MoveInternalReq:
      type: object
      required: [eventDate, householdId, newAddress]
      properties:
        eventDate:   { type: string, format: date }
        householdId: { type: string }
        newAddress:  { type: string }
        newAddressCode: { type: string }
    HouseholdChangeReq:
      type: object
      properties:
        operation:    { type: string, enum: [HEAD_CHANGE, SPLIT, MERGE] }
        householdId:  { type: string }
        newHeadResidentId: { type: string }
        eventDate:    { type: string, format: date }
        details:      { type: object }
    OfficialTxReq:
      type: object
      required: [residentId, reasonCode, eventDate, content]
      properties:
        residentId:  { type: string }
        reasonCode:  { type: string, enum: [OFFICIAL_WRITE, OFFICIAL_FIX, OFFICIAL_DELETE] }
        eventDate:   { type: string, format: date }
        legalBasis:  { type: string }
        content:     { type: string }
        attachments: { type: array, items: { type: string, format: uri } }
        approvalRoute: { type: array, items: { type: string, description: roleId } }
    ApprovalReq:
      type: object
      properties:
        action:   { type: string, enum: [APPROVE, REMAND, REJECT, CONDITIONAL] }
        comment:  { type: string }
    CancelReq:
      type: object
      required: [transactionId, reason]
      properties:
        transactionId: { type: string }
        reason:        { type: string }
...
    CertificateIssue:
      type: object
      properties:
        issueId:     { type: string }
        residentId:  { type: string }
        formId:      { type: string }
        copies:      { type: integer }
        fee:         { type: integer }
        verifyToken: { type: string }
        pdfUrl:      { type: string, format: uri }
        issuedAt:    { type: string, format: date-time }
        channel:     { type: string, enum: [WINDOW, CVS, ONLINE] }
...
    TransactionWithCertificate:
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            certificate: { $ref: '#/components/schemas/CertificateIssue' }
...